Add files via upload
Finished the important stuff. The program can now submit tests or pools of questions from an Excel file successfully. Had to make some minor tweaks to excelTest.py and question.py to get it to work. I will go through it later and clean it up as much as I can.

Left to do: user options